XRDDEV-237 refactored api key persistence
