--- src/proxy-ui-api/src/main/java/org/niis/xroad/restapi/repository/ApiKey2Repository.java
@@ -1,80 +1,0 @@
-/**
- * The MIT License
- * Copyright (c) 2018 Estonian Information System Authority (RIA),
- * Nordic Institute for Interoperability Solutions (NIIS), Population Register Centre (VRK)
- * Copyright (c) 2015-2017 Estonian Information System Authority (RIA), Population Register Centre (VRK)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package org.niis.xroad.restapi.repository;
-
-import lombok.extern.slf4j.Slf4j;
-import org.hibernate.Session;
-import org.niis.xroad.restapi.dao.ApiKeyDAOImpl;
-import org.niis.xroad.restapi.domain.ApiKeyType;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-
-import javax.persistence.EntityManager;
-
-import java.util.Collection;
-
-/**
- * Dummy in-memory repository for api keys
- */
-
-@Slf4j
-@Repository
-@Transactional
-public class ApiKey2Repository {
-
-    @Autowired
-    private EntityManager entityManager;
-
-    private Session getCurrentSession() {
-        return entityManager.unwrap(Session.class);
-    }
-
-    /**
-     * get all
-     * @return
-     */
-    public Collection<ApiKeyType> getApiKeys() {
-        ApiKeyDAOImpl apiKeyDAO = new ApiKeyDAOImpl();
-        return apiKeyDAO.findAll(getCurrentSession());
-    }
-
-    /**
-     * return one key
-     * @param id
-     */
-    public ApiKeyType getApiKey(long id) {
-        ApiKeyDAOImpl apiKeyDAO = new ApiKeyDAOImpl();
-        return apiKeyDAO.findById(getCurrentSession(), id);
-    }
-
-    /**
-     * save
-     */
-    public long insertApiKey(ApiKeyType apiKeyType) {
-        ApiKeyDAOImpl apiKeyDAO = new ApiKeyDAOImpl();
-        return apiKeyDAO.insert(getCurrentSession(), apiKeyType);
-    }
-}
